requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
bcrypt==5.0.0
bleach==6.3.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
contextlib2==21.6.0
cryptography==46.0.3
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.8
django-allauth==0.61.1
django-cors-headers==4.4.0
django-ninja==1.4.5
django-ninja-extra==0.30.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-yasg==1.21.11
email-validator==2.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==22.0.0
h11==0.16.0
httplib2==0.31.0
httptools==0.7.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
injector==0.22.0
mysql-connector-python==9.6.0
numpy==2.2.6
oauthlib==3.3.1
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==10.4.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.10
psycopg-binary==3.2.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pytesseract==0.3.13
pytest==9.0.1
pytest-django==4.11.1
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
pyzbar==0.1.9
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
webencodings==0.5.1
websockets==15.0.1
whitenoise==6.7.0
